pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autopptx"
version = "0.1.0"
description = "A Python tool for automated editing of PowerPoint templates using python-pptx"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

authors = [
  {name = "Chenzhe", email = "chenzhe0_0@163.com"},
]

maintainers = [
  {name = "Chenzhe", email = "chenzhe0_0@163.com"},
]

keywords = ["powerpoint", "automation", "presentation", "python-pptx", "template"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Office/Business :: Presentation",
  "Topic :: Software Development :: Libraries :: Python Modules"
]

dependencies = [
  "python-pptx>=1.0.2",
  "lxml>=5.3.1"
]

[project.urls]
Homepage = "https://github.com/chenzhex/AutoPPTX"
Documentation = "https://github.com/chenzhex/AutoPPTX#readme"
Repository = "https://github.com/chenzhex/AutoPPTX"
Issues = "https://github.com/chenzhex/AutoPPTX/issues"

[tool.setuptools.packages.find]
where = ["autopptx"]

[tool.setuptools]
include-package-data = true